Changed every reference to max data length and key variables to uint32 and bytes8

+++ document.sol
@@ -8,14 +8,14 @@ contract Document is DocumentAbstract {
   // bytes32 keyMap = 00001100000011000000110000011
   //  Where the 1s are the bytes where the key are
   function Document(byte[] _data, bytes21 head) {
-    for (uint64 i = 0; i < 21; i++) {
+    for (uint32 i = 0; i < 21; i++) {
         data.push(head[i]);
     }
     for (i = 0; i < _data.length - 4; i++) {
         data.push(_data[i + 4]);
     }
 
-    length = uint64(data.length);
+    length = uint32(data.length);
   }
 
   function getData() constant returns (byte[]) {
